# Add new `Performance/ArraySemiInfiniteRangeSlice` cop