@such 2020-11-16
`String#drop` and `String#take` don't exist but we can use `string[2..0]` and the Cop will complain (and auto-correct if asked to). Is there something else we can use or do we have to disable the Cop in that case?