@basex 2020-11-17
This cop is not safe. Example:

```ruby
".ext"[1..]
 => "ext" 
".ext".drop(1)
NoMethodError (undefined method `drop' for ".ext":String)
```