@marcandre 2020-11-23
Didn't anyone file a bug report on MRI for this? The solution here is not to optimize Ruby code but MRI...