@marcandre 2020-11-23
Actually, this is just an error in the benchmarking. The creation of the range `..2` and `90..` is what is taking extra time. This has nothing to do with semi-infinite ranges at all, or `drop`/`rake` being quicker. Note: to avoid creating a Range, on can use `Array#[index, length]`. In general though, `Array#[range]` often is the clearest.